backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
psycopg[binary]==3.2.9
psycopg-pool==3.2.6
redis==6.4.0
httpx==0.28.1
python-multipart==0.0.20
python-docx==1.2.0
openpyxl==3.1.5
pytest==8.4.2
